+++ module-client/src/main/java/org/module/client/businesslogicservice/statisticBLservice/CostAndIncomeBLService.java
@@ -13,14 +13,14 @@ public interface CostAndIncomeBLService {
 	 * 依赖：CostDataService.getAll 返回所有的付款单信息
 	 * @return
 	 */
-	public double getCost(double startTime,double endTime);
+	public double getCost(long startTime,long endTime);
 	/**
 	 * 前置：选择查看成本收益表
 	 * 后置：显示总收益
 	 * 依赖：ReceiptListService.getAll 返回所有的收款单
 	 * @return
 	 */
-	public double getIncome(double startTime,double endTime);
+	public double getIncome(long startTime,long endTime);
 	
 	/**
 	 * 前置：选择查看成本收益表
@@ -28,5 +28,5 @@ public interface CostAndIncomeBLService {
 	 * 依赖：getCost();getIncome();
 	 * @return
 	 */
-	public double getPureIncome(double startTime,double endTime);
+	public double getPureIncome(long startTime,long endTime);
 	}
